requirements.txt
Django~=4.2.6
av~=10.0.0
numpy~=1.26.0
torch~=2.1.0
requests~=2.31.0
djangorestframework~=3.14.0
transformers~=4.34.0
pip~=23.2.1
attrs~=22.2.0
wheel~=0.41.2
cryptography~=41.0.4
Pillow~=10.0.1
tornado~=6.3.3
Jinja2~=3.1.2
setuptools~=65.5.0
packaging~=23.1
pyparsing~=3.1.1
gunicorn~=21.2.0
